+++ tests/test_ext.py
@@ -27,12 +27,3 @@ class ExtensionTest(unittest.TestCase):
 
     def test_validate_environment_does_nothing_by_default(self):
         self.assertIsNone(self.ext.validate_environment())
-
-    def test_get_frontend_classes_returns_an_empty_list(self):
-        self.assertListEqual(self.ext.get_frontend_classes(), [])
-
-    def test_get_backend_classes_returns_an_empty_list(self):
-        self.assertListEqual(self.ext.get_backend_classes(), [])
-
-    def test_register_gstreamer_elements_does_nothing_by_default(self):
-        self.assertIsNone(self.ext.register_gstreamer_elements())